% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearbit-package.R
\docType{package}
\name{clearbit}
\alias{clearbit}
\alias{clearbit-package}
\title{...}
\description{
Clearbit provides an interface to perform queries on organizations and people.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/clearbit}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/clearbit/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
