% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{domain_to_org}
\alias{domain_to_org}
\title{The Company Name to Domain API lets you convert the exact name of a company
to a website domain, and a logo.}
\usage{
domain_to_org(domain, api_key = clearbit_api_key())
}
\arguments{
\item{domain}{name of the company to look for}

\item{api_key}{Clearbit API key. See \code{\link[=clearbit_api_key]{clearbit_api_key()}}}
}
\value{
data frame if found. \code{NULL} if not
}
\description{
The Company API allows you to look up a company by their domain. Alongside
the domain you may also provide any additional attributes you have about the
company, such as their company name or twitter handle. Including more detail
will help us be more accurate when searching.
}
