% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{name_to_domain}
\alias{name_to_domain}
\title{The Company Name to Domain API lets you convert the exact name of a company
to a website domain, and a logo.}
\usage{
name_to_domain(name, api_key = clearbit_api_key())
}
\arguments{
\item{name}{name of the company to look for}

\item{api_key}{Clearbit API key. See \code{\link[=clearbit_api_key]{clearbit_api_key()}}}
}
\value{
data frame containing \code{domain}, \code{logo}, and \code{name} if found. \code{NULL} if not
}
\description{
Clearbit matches based on exact company name, and returns the company with the most
website traffic. It’s worth noting, that since a company name is not a unique
identifier, you should be prepared for some inherent inaccuracy in the results.
}
